repaired duplicate column conflicts
When unnesting spectra columns, some common columns
were duplicated, including 'temperature', which is
representative of a thermal ramp in the spectra but
a DLS condition in the summary data.

This is also addressed in the util_get_spec_tbls.R
function which removed common columns 'created_at'
and 'updated_at'.