changed spectra ridgeline facet variable to 'well_id'
When multiple experiments were selected, i.e. a db query which
contained more than one 'uncle_experiment_set_id', the string for
well location could be used, thus 'well' alone was not sufficient
for generating the facets. This commit makes 'well_id' the only
facet variable, and adds a custom labeller to the 'ggridgeline'
function to make it clear to which experiment each well belongs.