removed the pre-built palette and instead rely on cbind_colors
The pre-built palette used in the method introduced in bd950a3 is
not a good way to go about reproducibley coloring wells to match
the ridgeline spectra. This is because the data is split on the
experiment_set_id as well as the color aesthetic variable and
therefore the palette is not always the correct length. Instead,
cbind_colors will be used to encode at an earlier step and then
the unique value of that color_hex variable following the split is
used to define the well color.